+++ src/main/java/com/onna/onnaback/global/oauth/handler/OAuth2LoginSuccessHandler.java
@@ -32,52 +32,9 @@ public class OAuth2LoginSuccessHandler extends SimpleUrlAuthenticationSuccessHan
     @Override
     public void onAuthenticationSuccess(HttpServletRequest request, HttpServletResponse response, Authentication authentication) throws IOException, ServletException {
         log.info("OAuth2 Login 성공!");
-        try {
-            CustomOAuth2User oAuth2User = (CustomOAuth2User) authentication.getPrincipal();
-
-            // User의 Role이 GUEST일 경우 처음 요청한 회원이므로 회원가입 페이지로 리다이렉트
-            if(oAuth2User.getRole() == Role.GUEST) {
-//                String accessToken = jwtService.createAccessToken(oAuth2User.getEmail());
-//                response.addHeader(jwtService.getAccessHeader(), "Bearer " + accessToken);
-//                response.sendRedirect("oauth2/sign-up"); // 프론트의 회원가입 추가 정보 입력 폼으로 리다이렉트
-//
-//                jwtService.sendAccessAndRefreshToken(response, accessToken, null);
-                String targetUrl = UriComponentsBuilder.fromUriString(redirectUrl)
-                        .queryParam("email", (String) oAuth2User.getAttribute("email"))
-                        .build()
-                        .encode(StandardCharsets.UTF_8)
-                        .toUriString();
-                // 회원가입 페이지로 리다이렉트 시킨다.
-                getRedirectStrategy().sendRedirect(request, response, targetUrl);
-
-            } else {
-//                loginSuccess(response, oAuth2User); // 로그인에 성공한 경우 access, refresh 토큰 생성
-                // accessToken을 쿼리스트링에 담는 url을 만들어준다.
-                System.err.println(redirectUrl);
-                String targetUrl = UriComponentsBuilder.fromUriString(redirectUrl)
-                        .queryParam("accessToken", jwtService.createAccessToken(oAuth2User.getEmail()))
-                        .build()
-                        .encode(StandardCharsets.UTF_8)
-                        .toUriString();
-                log.info("redirect 준비");
-                // 로그인 확인 페이지로 리다이렉트 시킨다.
-                getRedirectStrategy().sendRedirect(request, response, targetUrl);
-            }
-        } catch (Exception e) {
-            throw e;
-        }
+        CustomOAuth2User oAuth2User = (CustomOAuth2User) authentication.getPrincipal();
+        response.getWriter().write( "Bearer "+jwtService.createAccessToken(oAuth2User.getEmail()));
 
     }
 
-
-
-    private void loginSuccess(HttpServletResponse response, CustomOAuth2User oAuth2User) throws IOException {
-        String accessToken = jwtService.createAccessToken(oAuth2User.getEmail());
-        String refreshToken = jwtService.createRefreshToken();
-        response.addHeader(jwtService.getAccessHeader(), "Bearer " + accessToken);
-        response.addHeader(jwtService.getRefreshHeader(), "Bearer " + refreshToken);
-
-        jwtService.sendAccessAndRefreshToken(response, accessToken, refreshToken);
-        jwtService.updateRefreshToken(oAuth2User.getEmail(), refreshToken);
-    }
 }
